🦄refactor(CadViewer): removing legacy - stage 3
