+++ src/app/cad-viewer/cad-data/cad-types.ts
@@ -1,6 +1,6 @@
 export type CadType = "LINE" | "MTEXT" | "DIMENSION" | "ARC" | "CIRCLE" | "HATCH";
 
-export const cadTypes = ["LINE", "MTEXT", "DIMENSION", "ARC", "CIRCLE", "HATCH"];
+export const cadTypes: CadType[] = ["LINE", "MTEXT", "DIMENSION", "ARC", "CIRCLE", "HATCH"];
 
 export type CadTypeKey = "line" | "mtext" | "dimension" | "arc" | "circle" | "hatch";
