+++ src/app/cad-viewer/cad-data/utils.ts
@@ -1,5 +1,5 @@
 import {Point} from "@src/app/utils";
-import {CadLine} from "./cad-entity/cad-line";
+import {CadLine} from "./cad-entity";
 
 export function getVectorFromArray(data: number[] | Point, defaultVal = new Point()) {
 	if (data instanceof Point) {
