+++ src/app/components/cad-console/cad-console.component.ts
@@ -1,25 +1,24 @@
-import {Component, OnInit, ViewChild, ElementRef, Input, Injector, Output, EventEmitter, OnDestroy} from "@angular/core";
-import {differenceWith} from "lodash";
-import {timeout, removeCadGongshi, Collection, addCadGongshi, Command, getDPI, session} from "@src/app/app.common";
+import {trigger, transition, style, animate} from "@angular/animations";
+import {Component, OnInit, OnDestroy, ViewChild, ElementRef, Input, Output, EventEmitter, Injector} from "@angular/core";
 import {MatSnackBar} from "@angular/material/snack-bar";
-import {CadViewer} from "@src/app/cad-viewer/cad-viewer";
-import {CadData, CadOption, CadBaseLine, CadJointPoint} from "@src/app/cad-viewer/cad-data/cad-data";
-import {openMessageDialog, MessageComponent} from "../message/message.component";
-import {validateLines} from "@src/app/cad-viewer/cad-data/cad-lines";
-import {MenuComponent} from "../menu/menu.component";
-import {CurrCadsAction, CadStatusAction, LoadingAction} from "@src/app/store/actions";
-import {openCadListDialog} from "../cad-list/cad-list.component";
-import {getCadStatus, getCommand, getCurrCads, getCurrCadsData} from "@src/app/store/selectors";
-import {highlight} from "highlight.js";
-import {CadDimension} from "@src/app/cad-viewer/cad-data/cad-entity/cad-dimension";
-import {createPdf} from "pdfmake/build/pdfmake";
-import {Angle, Line, Point} from "@app/utils";
-import {CadArc} from "@src/app/cad-viewer/cad-data/cad-entity/cad-arc";
-import {animate, style, transition, trigger} from "@angular/animations";
 import {ActivatedRoute} from "@angular/router";
-import Color from "color";
+import {Command, timeout, addCadGongshi, removeCadGongshi, getDPI, session, Collection} from "@src/app/app.common";
+import {CadData, CadOption, CadBaseLine, CadJointPoint} from "@src/app/cad-viewer/cad-data/cad-data";
+import {CadArc, CadDimension} from "@src/app/cad-viewer/cad-data/cad-entity";
+import {validateLines} from "@src/app/cad-viewer/cad-data/cad-lines";
+import {CadViewer} from "@src/app/cad-viewer/cad-viewer";
+import {CurrCadsAction, CadStatusAction, LoadingAction} from "@src/app/store/actions";
+import {getCommand, getCurrCads, getCurrCadsData, getCadStatus} from "@src/app/store/selectors";
+import {Angle, Line, Point} from "@src/app/utils";
+import {MatrixAlias, Matrix} from "@svgdotjs/svg.js";
+import {differenceWith} from "lodash";
 import {v4} from "uuid";
-import {Matrix, MatrixAlias} from "@svgdotjs/svg.js";
+import {openCadListDialog} from "../cad-list/cad-list.component";
+import {MenuComponent} from "../menu/menu.component";
+import {openMessageDialog, MessageComponent} from "../message/message.component";
+import {highlight} from "highlight.js";
+import Color from "color";
+import {createPdf} from "pdfmake/build/pdfmake";
 
 const getList = (content: string[]) => {
 	return `<ul>${content.map((v) => `<li>${v}</li>`).join("")}</ul>`;
