+++ src/app/utils/index.ts
@@ -3,7 +3,7 @@ import {Line} from "./geometry/line";
 import {Rectangle} from "./geometry/rectangle";
 import {Arc} from "./geometry/arc";
 import {Angle} from "./geometry/angle";
-import {RGB2Index, index2RGB, getColorLightness} from "./color";
+import {RGB2Index, index2RGB} from "./color";
 import {dataURLtoBlob, RSAEncrypt} from "./misc";
 import {SessionStorage, LocalStorage} from "./storage";
 import {DEFAULT_TOLERANCE} from "./geometry/constants";
@@ -19,7 +19,6 @@ export {
 	RSAEncrypt,
 	RGB2Index,
 	index2RGB,
-	getColorLightness,
 	SessionStorage,
 	LocalStorage
 };
