+++ tsconfig.app.json
@@ -9,6 +9,6 @@
     "src/polyfills.ts"
   ],
   "include": [
-    "src/**/*.d.ts"
+    "src/**/*.d.ts", "src/app/cad-viewer/cad-data/index.ts"
   ]
 }
