+++ tsconfig.spec.json
@@ -13,6 +13,6 @@
   ],
   "include": [
     "src/**/*.spec.ts",
-    "src/**/*.d.ts"
+    "src/**/*.d.ts", "src/app/cad-viewer/cad-data/index.ts"
   ]
 }
